+++ test/provider_consumer_test.dart
@@ -176,7 +176,7 @@ void main() {
       Vector({}, {'rsp_data': 9}),
     ];
 
-    final sv = mod.generateSynth();
+    final sv = mod.dumpSystemVerilog().output;
 
     expect(
         sv,