+++ test/provider_consumer_w_modify_test.dart
@@ -147,7 +147,7 @@ void main() {
       Vector({}, {'rsp_data': 9}),
     ];
 
-    final sv = mod.generateSynth();
+    final sv = mod.dumpSystemVerilog().output;
 
     expect(
         sv,